+++ ffn_dl/__init__.py
@@ -2,7 +2,7 @@
 
 import logging as _logging
 
-__version__ = "2.2.31"
+__version__ = "2.3.0"
 
 _logger = _logging.getLogger(__name__)